Remove `Color` module and move functionality into Pixel
