+++ lib/ex_png/pixel.ex
@@ -3,6 +3,8 @@ defmodule ExPng.Pixel do
   Represents a single pixel in an image, storing red, green, blue and alpha
   values, or an index when part of a paletted image.
   """
+  use ExPng.Constants
+  use Bitwise
 
   @type rgba_value :: 0..255
   @type t :: %__MODULE__{
@@ -74,6 +76,45 @@ defmodule ExPng.Pixel do
   def black_or_white?(%__MODULE__{r: 0, g: 0, b: 0, a: 255}), do: true
   def black_or_white?(%__MODULE__{r: 255, g: 255, b: 255, a: 255}), do: true
   def black_or_white?(_), do: false
+
+  def pixel_bytesize(%ExPng.RawData{header_chunk: header}) do
+    pixel_bytesize(header.color_mode, header.bit_depth)
+  end
+
+  def pixel_bytesize(color_mode, bit_depth \\ 8) do
+    color_mode
+    |> pixel_bitsize(bit_depth)
+    |> to_bytesize()
+  end
+
+  def line_bytesize(%ExPng.RawData{header_chunk: header}) do
+    line_bytesize(header.color_mode, header.bit_depth, header.width)
+  end
+
+  def line_bytesize(color_mode, bit_depth, width) do
+    color_mode
+    |> pixel_bitsize(bit_depth)
+    |> Kernel.*(width)
+    |> to_bytesize()
+  end
+
+  defp pixel_bitsize(color_mode, bit_depth) do
+    color_mode
+    |> channels_for_color_mode()
+    |> Kernel.*(bit_depth)
+  end
+
+  defp channels_for_color_mode(@grayscale), do: 1
+  defp channels_for_color_mode(@truecolor), do: 3
+  defp channels_for_color_mode(@indexed), do: 1
+  defp channels_for_color_mode(@grayscale_alpha), do: 2
+  defp channels_for_color_mode(@truecolor_alpha), do: 4
+
+  defp to_bytesize(x) do
+    x
+    |> Kernel.+(7)
+    |> Bitwise.>>>(3)
+  end
 end
 
 defimpl Inspect, for: ExPng.Pixel do
